requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
blurhash==1.1.4
    # via mastodon-py
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
decorator==5.2.1
    # via mastodon-py
idna==3.10
    # via requests
mastodon-py==2.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via mastodon-py
python-dotenv==1.0.1
    # via -r requirements.in
python-magic==0.4.27
    # via mastodon-py
requests==2.32.3
    # via
    #   -r requirements.in
    #   mastodon-py
six==1.17.0
    # via python-dateutil
urllib3==2.3.0
    # via requests
